Fix: skip install if HPGMG or PETSc already there
- Also, don't run BLIS test - just print a log message about how to run it (similar to HPGMG)

+++ cookbooks/emulab-petsc/recipes/default.rb
@@ -17,6 +17,8 @@ bash 'PETSC: configure, make, install' do
     ./configure --with-debugging=0 --prefix="#{node['petsc']['root_level_dir']}/petsc" --download-f2cblaslapack=1 --with-shared-libraries=0 2>&1 > .configure.log 
     make MAKE_NP=#{node["cpu"]["total"]} all 2>&1 > .make.all.log
     make install 2>&1 > .make.install.log
+    touch "#{node['petsc']['root_level_dir']}/petsc/.completed"
   EOH
   cwd "#{node['petsc']['root_level_dir']}/petsc-src"
+  not_if { File.exists?("#{node['petsc']['root_level_dir']}/petsc/.completed") }
 end
